Refactor config and exploit code into appropriate packages
