+++ client/cmd/root.go
@@ -31,7 +31,7 @@ func Execute() {
 func init() {
 	RootCmd.AddCommand(ConfigCmd)
 	RootCmd.PersistentFlags().BoolVarP(&config.ArgsAttackInstance.Debug, "debug", "D", false, "Enable debug logging")
-	RootCmd.PersistentFlags().BoolVarP(&config.NoTUI, "no-tui", "N", false, "Disable TUI mode")
+	RootCmd.PersistentFlags().BoolVarP(&config.UseTUI, "no-tui", "N", false, "Disable TUI mode")
 	RootCmd.PersistentFlags().BoolVarP(&config.UseBanner, "no-banner", "B", false, "Remove banner on startup")
 
 	RootCmd.PersistentPreRun = func(cmd *cobra.Command, args []string) {
